# Perf: Mixed-precision p_cmp logits (on workspace-presize)